Move compress() and decompress() to dedicated library crate
