+++ mozlz4/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "mozlz4-bin"
+name = "mozlz4"
 version = "0.1.0"
+edition = "2021"
 authors = ["Justin Wong <jusw85@hotmail.com>"]
 
 [dependencies]
-error-chain = "0.11"
-clap = "2.30"
 byteorder = "1"
-mozlz4 = { path = "../mozlz4" }
+error-chain = "0.11"
+mozlz4-sys = { path = "../mozlz4-sys" }
